+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity.AdminRegActivity"></activity>
+        <activity android:name=".AuthActivity.AdminRegActivity"></activity>
         <activity android:name=".MainActivity.UserHomeActivity" />
         <activity android:name=".MainActivity.AdminHomeActivity" />
         <activity android:name=".AuthActivity.LoginActivity" />
