+++ app/src/main/java/com/rrr/swift/MainActivity/AdminHomeActivity.java
@@ -10,9 +10,9 @@ import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.Button;
 import android.widget.Toast;
 
+import com.rrr.swift.AuthActivity.AdminRegActivity;
 import com.rrr.swift.HomeActivity;
 import com.rrr.swift.LocationActivities.LocationActivity;
 import com.rrr.swift.R;
