+++ app/src/main/res/layout/activity_admin_reg.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity.AdminRegActivity">
+    tools:context=".AuthActivity.AdminRegActivity">
 
 </android.support.constraint.ConstraintLayout>
